many: move AA profile operations to sandbox/apparmor
We might need to trigger profile reloading from other parts of snapd, to
support updating the HOMEDIRS AppArmor tunable.
