+++ interfaces/apparmor/backend.go
@@ -65,6 +65,8 @@ var (
 	isRootWritableOverlay = osutil.IsRootWritableOverlay
 	kernelFeatures        = apparmor_sandbox.KernelFeatures
 	parserFeatures        = apparmor_sandbox.ParserFeatures
+	loadProfiles          = apparmor_sandbox.LoadProfiles
+	unloadProfiles        = apparmor_sandbox.UnloadProfiles
 
 	// make sure that apparmor profile fulfills the late discarding backend
 	// interface
@@ -207,9 +209,9 @@ func (b *Backend) Initialize(opts *interfaces.SecurityBackendOptions) error {
 		return nil
 	}
 
-	aaFlags := skipReadCache
+	aaFlags := apparmor_sandbox.SkipReadCache
 	if b.preseed {
-		aaFlags |= skipKernelLoad
+		aaFlags |= apparmor_sandbox.SkipKernelLoad
 	}
 
 	if err := loadProfiles([]string{profilePath}, apparmor_sandbox.SystemCacheDir, aaFlags); err != nil {
@@ -333,9 +335,9 @@ func (b *Backend) setupSnapConfineReexec(info *snap.Info) error {
 		pathnames[i] = filepath.Join(dir, profile)
 	}
 
-	var aaFlags aaParserFlags
+	var aaFlags apparmor_sandbox.AaParserFlags
 	if b.preseed {
-		aaFlags = skipKernelLoad
+		aaFlags = apparmor_sandbox.SkipKernelLoad
 	}
 	errReload := loadProfiles(pathnames, cache, aaFlags)
 	errUnload := unloadProfiles(removed, cache)
@@ -499,9 +501,9 @@ func (b *Backend) Setup(snapInfo *snap.Info, opts interfaces.ConfinementOptions,
 	// work despite time being wrong (e.g. in the past). For more details see
 	// https://forum.snapcraft.io/t/apparmor-profile-caching/1268/18
 	var errReloadChanged error
-	aaFlags := skipReadCache
+	aaFlags := apparmor_sandbox.SkipReadCache
 	if b.preseed {
-		aaFlags |= skipKernelLoad
+		aaFlags |= apparmor_sandbox.SkipKernelLoad
 	}
 	timings.Run(tm, "load-profiles[changed]", fmt.Sprintf("load changed security profiles of snap %q", snapInfo.InstanceName()), func(nesttm timings.Measurer) {
 		errReloadChanged = loadProfiles(prof.changed, apparmor_sandbox.CacheDir, aaFlags)
@@ -513,7 +515,7 @@ func (b *Backend) Setup(snapInfo *snap.Info, opts interfaces.ConfinementOptions,
 	var errReloadOther error
 	aaFlags = 0
 	if b.preseed {
-		aaFlags |= skipKernelLoad
+		aaFlags |= apparmor_sandbox.SkipKernelLoad
 	}
 	timings.Run(tm, "load-profiles[unchanged]", fmt.Sprintf("load unchanged security profiles of snap %q", snapInfo.InstanceName()), func(nesttm timings.Measurer) {
 		errReloadOther = loadProfiles(prof.unchanged, apparmor_sandbox.CacheDir, aaFlags)
@@ -551,18 +553,18 @@ func (b *Backend) SetupMany(snaps []*snap.Info, confinement func(snapName string
 	}
 
 	if !fallback {
-		aaFlags := skipReadCache | conserveCPU
+		aaFlags := apparmor_sandbox.SkipReadCache | apparmor_sandbox.ConserveCPU
 		if b.preseed {
-			aaFlags |= skipKernelLoad
+			aaFlags |= apparmor_sandbox.SkipKernelLoad
 		}
 		var errReloadChanged error
 		timings.Run(tm, "load-profiles[changed-many]", fmt.Sprintf("load changed security profiles of %d snaps", len(snaps)), func(nesttm timings.Measurer) {
 			errReloadChanged = loadProfiles(allChangedPaths, apparmor_sandbox.CacheDir, aaFlags)
 		})
 
-		aaFlags = conserveCPU
+		aaFlags = apparmor_sandbox.ConserveCPU
 		if b.preseed {
-			aaFlags |= skipKernelLoad
+			aaFlags |= apparmor_sandbox.SkipKernelLoad
 		}
 		var errReloadOther error
 		timings.Run(tm, "load-profiles[unchanged-many]", fmt.Sprintf("load unchanged security profiles %d snaps", len(snaps)), func(nesttm timings.Measurer) {
@@ -620,7 +622,7 @@ func (b *Backend) RemoveLate(snapName string, rev snap.Revision, typ snap.Type) 
 
 	globs := []string{snapConfineProfileName(snapName, rev)}
 	_, removed, errEnsure := osutil.EnsureDirStateGlobs(dirs.SnapAppArmorDir, globs, nil)
-	// XXX: unloadProfiles() does not unload profiles from the kernel, but
+	// XXX: UnloadProfiles() does not unload profiles from the kernel, but
 	// only removes profiles from the cache
 	// always try to unload the affected profile
 	errUnload := unloadProfiles(removed, apparmor_sandbox.CacheDir)
